Rotating only image, changing styles on different drag states
